moving CreateNewPlaylistModal component to its own file and am using it where it is demanded
