+++ src/components/layout/Sidebar.tsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import { useCookies } from "react-cookie";
 import * as ScrollArea from "@radix-ui/react-scroll-area";
 import {
   MagnifyingGlass,
@@ -9,13 +8,10 @@ import {
   Plus,
   ArrowRight,
 } from "@phosphor-icons/react";
-import { PlaylistItem } from "../playlist/PlaylistItem";
-import { Modal } from "../reusable/Modal";
-import { PlaylistModalContent } from "../playlist/PlaytlistModalContent";
-import { useUserDataContext } from "../../contexts/UserDataContext";
-import { type Playlist as PlaylistInfo } from "../../services/playlistServices";
-
-const SERVER_URL = import.meta.env.VITE_SERVER_URL;
+import { PlaylistItem } from "@components/playlist/PlaylistItem";
+import { useUserDataContext } from "@contexts/UserDataContext";
+import { CreateNewPlaylistModal } from "@components/playlist/CreateNewPlaylistModal";
+import { type Playlist as PlaylistInfo } from "@services/playlistServices";
 
 export const Sidebar = () => {
   const [selected, setSelected] = useState<"home" | "search">("home");
@@ -92,56 +88,5 @@ export const Sidebar = () => {
         <CreateNewPlaylistModal handleClose={() => setShowModal(false)} />
       )}
     </>
-  );
-};
-
-const CreateNewPlaylistModal = ({
-  handleClose,
-}: {
-  handleClose: () => void;
-}) => {
-  const [cookies] = useCookies(["user_jwt"]);
-  const [playlistInfo, setPlaylistData] = useState<{
-    name?: string;
-    description?: string;
-  }>();
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ) => {
-    const fieldName = e.target.name;
-    const fieldData = e.target.value;
-
-    setPlaylistData((prev) => ({ ...prev, [fieldName]: fieldData }));
-  };
-
-  const handleSaveEdit = async () => {
-    try {
-      if (playlistInfo?.name) {
-        console.log("sending new playlist data", {
-          ...playlistInfo,
-        });
-
-        const createPlaylistRes = await fetch(`${SERVER_URL}/playlist/create`, {
-          method: "POST",
-          headers: { Authorization: `Bearer ${cookies.user_jwt}` },
-          body: JSON.stringify({
-            name: playlistInfo?.name,
-            description: playlistInfo?.description,
-          }),
-        });
-
-        return createPlaylistRes;
-      }
-    } catch (error) {
-      console.error("Erro while trying to submit the edit.", error);
-    }
-  };
-  return (
-    <Modal handleClose={handleClose} title="Criar Playlist">
-      <PlaylistModalContent
-        handleChange={handleChange}
-        handleSaveEdit={handleSaveEdit}
-      />
-    </Modal>
   );
 };
